+++ src/main/java/JYQ/ShortestDistancePathFinder/ShortestPathFinder.java
@@ -7,13 +7,13 @@ import java.util.*;
 
 public class ShortestPathFinder {
        private DistanceGraph Graph;
-       private int[] DistanceTo;
+       private double[] DistanceTo;
        private int[] EdgeTo;
        private boolean[] Removed;
        //private PriorityQueue<Integer> NearestVertices;
        public ShortestPathFinder(DistanceGraph graph) {
            Graph = graph;
-           DistanceTo = new int[graph.getVertices()];
+           DistanceTo = new double[graph.getVertices()];
            EdgeTo = new int[graph.getVertices()];
            Removed = new boolean[graph.getVertices()];
            /*Comparator<Integer> comparator = (Integer o1, Integer o2) ->{
@@ -54,7 +54,7 @@ public class ShortestPathFinder {
                for (int i = 0 ; i < Adjacent.length;i ++) {
                    WeigtedEgde weigtedEgde = Adjacent[i];
                    if (weigtedEgde != null) {
-                       int NewDistance = weigtedEgde.getWeight() + DistanceTo[bestNode];
+                       double NewDistance = weigtedEgde.getWeight() + DistanceTo[bestNode];
                        int CurrentAdjacent = weigtedEgde.getAdjacentVertexNumber();
                        if (DistanceTo[CurrentAdjacent] == -1 || DistanceTo[CurrentAdjacent] > NewDistance) {
                            DistanceTo[CurrentAdjacent] = NewDistance;
